Big changes v2
Modified the whole using namespaces in the project to make it look cleaner.
Also implemented the DataDragon almost completely
Now you can save the responses, if they're json files they will be saved as jsons, if they're png or jpg they will be saved as an image with the Media class.